requirements.txt
matplotlib~=3.3.3
keras~=2.4.3
joblib~=1.0.0
tensorflow-gpu~=2.4.0